+++ src/es_sys_scripts.rs
@@ -1,6 +1,6 @@
-use crate::esruntimewrapper::EsRuntimeWrapper;
+use crate::esruntime::EsRuntime;
 
-pub(crate) fn init_es(rt: &EsRuntimeWrapper) {
+pub(crate) fn init_es(rt: &EsRuntime) {
     init_file(
         rt,
         "es_sys_scripts/es_01_core.es",
@@ -8,7 +8,7 @@ pub(crate) fn init_es(rt: &EsRuntimeWrapper) {
     );
 }
 
-fn init_file(runtime: &EsRuntimeWrapper, file_name: &str, es_code: &str) {
+fn init_file(runtime: &EsRuntime, file_name: &str, es_code: &str) {
     let init_res = runtime.eval_void_sync(es_code, file_name);
     if init_res.is_err() {
         let esei = init_res.err().unwrap();